Перенесет render в отдельный файл и поправит сслыки import в main и abstract-component
