+++ src/utils.js
@@ -21,25 +21,3 @@ export const dateTimeConverter = (isDateShowing, dueDate) => {
   };
 };
 
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const renderSection = (placeInDocument, section, position) => {
-  switch (position) {
-    case RenderPosition.AFTERBEGIN:
-      placeInDocument.prepend(section);
-      break;
-    case RenderPosition.BEFOREEND:
-      placeInDocument.append(section);
-      break;
-  }
-};
